test: skip filemap heap test under race detector
We seem to be hitting a Go bug when it reports the data race from the
same goroutine in the stdlib deflate package.
